// src/entities/Country/model/types/CountrySchema.ts
export enum Country {
  Belarus = "Belarus",
  Russia = "Russia",
  UnitedStates = "United States",
  Germany = "Germany",
  Japan = "Japan",
  France = "France",
  UnitedKingdom = "United Kingdom",
  Canada = "Canada",
  Australia = "Australia",
  Sweden = "Sweden",
  Switzerland = "Switzerland",
  Netherlands = "Netherlands",
}

export interface CountrySchema {
  country: Country;
}